Mission 6.1.0 : menu move to fragmentfrom activity & viewmodel variable modification
