+++ app/src/main/java/com/survivalcoding/todolist/ui/main/EditFragment.kt
@@ -9,7 +9,6 @@ import android.view.inputmethod.InputMethodManager
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.survivalcoding.todolist.databinding.FragmentEditBinding
-import com.survivalcoding.todolist.domain.entity.Todo
 
 
 class EditFragment : Fragment() {
@@ -17,12 +16,18 @@ class EditFragment : Fragment() {
     private val binding get() = _binding!!
     private val mainViewModel by activityViewModels<MainViewModel>()
 
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        //setHasOptionsMenu(false)
+    }
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
         _binding = FragmentEditBinding.inflate(inflater, container, false)
+        setHasOptionsMenu(true)
         return binding.root
     }
 
@@ -34,30 +39,14 @@ class EditFragment : Fragment() {
             requireContext().getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
         manager.showSoftInput(binding.addEditText, 0)
 
-        //뭔가 지저분한 느낌
-        var isUpdate = false
-        var id: Long = -1
-        mainViewModel.isUpdate.value?.let { _isUpdate ->
-            if (_isUpdate) {
-                isUpdate = _isUpdate
-                mainViewModel.todoNeedChanged.value?.let {
-                    binding.addEditText.setText(it.title)
-                    id = it.id
-                }
-            }
+        mainViewModel.todoNeedChanged.observe(this) {
+            it?.let { binding.addEditText.setText(it.title) }
         }
 
         binding.addButton.setOnClickListener {
             val inputTodoName = binding.addEditText.text.toString().trim()
             if (inputTodoName.isEmpty()) return@setOnClickListener
-            if (isUpdate) {
-                mainViewModel.updateTodo(inputTodoName, id)
-                //todo 더 우아하게 할 수 있지 않을까?
-                mainViewModel.isUpdate.value = false
-                mainViewModel.todoNeedChanged.value = null
-            } else {
-                mainViewModel.addTodo(inputTodoName)
-            }
+            mainViewModel.upsertTodo(inputTodoName)
             parentFragmentManager.popBackStack()
         }
